requirements.txt
streamlit==1.25.0
pymupdf==1.22.5
python-docx==0.8.12
pytesseract==0.3.12
google-generativeai==0.1.0
python-dotenv==1.0.0

